fix #84

+++ common/buildcraft/compat/carpentersblocks/SchematicCBGate.java
@@ -2,8 +2,8 @@ package buildcraft.compat.carpentersblocks;
 
 public class SchematicCBGate extends SchematicCBRotated {
 	@Override
-	protected short fixMetadata(short m) {
-		short newM = (short) (m & ~0x20);
+	protected int fixMetadata(int m) {
+		int newM = m & ~0x20;
 		if ((m & 0x20) == 0) {
 			newM |= 0x20;
 		}
