+++ common/buildcraft/compat/carpentersblocks/SchematicCBRotatedTwo.java
@@ -2,7 +2,7 @@ package buildcraft.compat.carpentersblocks;
 
 public class SchematicCBRotatedTwo extends SchematicCBRotated {
 	@Override
-	protected short fixMetadata(short m) {
-		return (short) (m ^ 1);
+	protected int fixMetadata(int m) {
+		return (m ^ 1);
 	}
 }
