+++ common/buildcraft/compat/carpentersblocks/SchematicCBSafe.java
@@ -4,7 +4,7 @@ public class SchematicCBSafe extends SchematicCBRotated {
 	private static final short[] rotMatrix = {1, 2, 3, 0};
 
 	@Override
-	protected short fixMetadata(short m) {
-		return (short) ((m & ~0x3) | rotMatrix[m & 0x3]);
+	protected int fixMetadata(int m) {
+		return (m & ~0x3) | rotMatrix[m & 0x3];
 	}
 }
