+++ common/buildcraft/compat/carpentersblocks/SchematicCBTorch.java
@@ -2,7 +2,7 @@ package buildcraft.compat.carpentersblocks;
 
 public class SchematicCBTorch extends SchematicCBRotated {
 	@Override
-	protected short fixMetadata(short m) {
-		return (short) ((m & (~0x7)) | shiftMatrix[m & 0x7]);
+	protected int fixMetadata(int m) {
+		return ((m & (~0x7)) | shiftMatrix[m & 0x7]);
 	}
 }
